+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-## v0.6.40
+## v0.7.1
 
 - Add `HI` token to resolver list
+
+## v0.7.0
+
+### Breaking changes
+
+- Remove `setReverse(uint256)` function from `UNSRegistry`
+
+### Non-breaking changes
+
+- Remove `unlockWithRecords(address,uint256,string[],string[],bool)` function from `UNSRegistry`
 
 ## v0.6.39